Add batch_circle_nms and modify CenterHead

+++ autoware_ml/models/detection3d/dataclasses/targets.py
@@ -2,7 +2,7 @@
 Module to save encoded targets for a 3D detection head.
 """
 
-from jaxtyping import Float32, Int32, Bool
+from jaxtyping import Float32, Int64, Bool
 from pydantic import BaseModel, ConfigDict
 
 import torch
@@ -25,5 +25,5 @@ class CenterHeadTargets(BaseModel):
     # 8 (center_x, center_y, center_z, length, width, height, sin(heading), cos(heading)) if not velocity else
     # 10 (center_x, center_y, center_z, length, width, height, heading, velocity_x, velocity_y)
     reg_targets: Float32[torch.Tensor, "batch_size max_num_boxes num_reg_targets"]
-    reg_indices: Int32[torch.Tensor, "batch_size max_num_boxes"]
+    reg_indices: Int64[torch.Tensor, "batch_size max_num_boxes"]
     valid_masks: Bool[torch.Tensor, "batch_size max_num_boxes"]
